+++ x-pack/plugins/security_solution/server/lib/entity_analytics/asset_criticality/routes/check_and_init_asset_criticality_resources.ts
